+++ inference/python_api_test/test_class_model/test_pcpvt_base_trt_fp16.py
@@ -44,24 +44,8 @@ def test_config():
     test_suite.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
     test_suite.config_test()
 
-
 @pytest.mark.server
-@pytest.mark.config_disablegpu_memory
-def test_disable_gpu():
-    """
-    test no gpu resources occupied after disable gpu
-    """
-    check_model_exist()
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
-    batch_size = 1
-    fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
-    input_data_dict = {"x": fake_input}
-    test_suite.disable_gpu_test(input_data_dict)
-
-
-@pytest.mark.server
-@pytest.mark.trt_fp16_more_bz_precision
+@pytest.mark.trt_fp16
 def test_trt_fp16_more_bz():
     """
     compared trt fp16 batch_size=1-10 pcpvt_base outputs with true val
@@ -95,7 +79,7 @@ def test_trt_fp16_more_bz():
 
 
 @pytest.mark.jetson
-@pytest.mark.trt_fp16_more_bz_precision
+@pytest.mark.trt_fp16
 def test_jetson_trt_fp16_more_bz():
     """
     compared trt fp16 batch_size=1-10 pcpvt_base outputs with true val
@@ -104,7 +88,7 @@ def test_jetson_trt_fp16_more_bz():
 
     file_path = "./pcpvt_base"
     images_size = 224
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(
@@ -129,7 +113,7 @@ def test_jetson_trt_fp16_more_bz():
 
 
 @pytest.mark.server
-@pytest.mark.trt_fp16_multi_thread_bz1_precision
+@pytest.mark.trt_fp16_multi_thread
 def test_trt_fp16_bz1_multi_thread():
     """
     compared trt fp16 batch_size=1 pcpvt_base multi_thread outputs with true val
