+++ inference/python_api_test/test_class_model/test_resnet50_gpu.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # encoding=utf-8 vi:ts=4:sw=4:expandtab:ft=python
 """
-test pcpvt_base model
+test resnet50 model
 """
 
 import os
@@ -15,7 +15,7 @@ import numpy as np
 
 # pylint: disable=wrong-import-position
 sys.path.append("..")
-from test_case import InferenceTest
+from test_case import InferenceTest, clip_model_extra_op
 
 # pylint: enable=wrong-import-position
 
@@ -24,12 +24,13 @@ def check_model_exist():
     """
     check model exist
     """
-    tnt_small_url = "https://paddle-qa.bj.bcebos.com/inference_model/2.1.1/class/pcpvt_base.tgz"
-    if not os.path.exists("./pcpvt_base/inference.pdiparams"):
-        wget.download(tnt_small_url, out="./")
-        tar = tarfile.open("pcpvt_base.tgz")
+    resnet50_url = "https://paddle-qa.bj.bcebos.com/inference_model/2.0/class/resnet50.tgz"
+    if not os.path.exists("./resnet50/inference.pdiparams"):
+        wget.download(resnet50_url, out="./")
+        tar = tarfile.open("resnet50.tgz")
         tar.extractall()
         tar.close()
+        clip_model_extra_op(path_prefix="./resnet50/inference", output_model_path="./resnet50/inference")
 
 
 @pytest.mark.server
@@ -41,7 +42,7 @@ def test_config():
     """
     check_model_exist()
     test_suite = InferenceTest()
-    test_suite.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
+    test_suite.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
     test_suite.config_test()
 
 
@@ -53,99 +54,64 @@ def test_disable_gpu():
     """
     check_model_exist()
     test_suite = InferenceTest()
-    test_suite.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
+    test_suite.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
     batch_size = 1
     fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
-    input_data_dict = {"x": fake_input}
+    input_data_dict = {"inputs": fake_input}
     test_suite.disable_gpu_test(input_data_dict)
 
 
 @pytest.mark.server
-@pytest.mark.mkldnn_bz1_precision
-def test_mkldnn():
-    """
-    compared mkldnn pcpvt_base outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./pcpvt_base"
-    images_size = 224
-    batch_size = 1
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
-    images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-    fake_input = np.array(images_list[0:batch_size]).astype("float32")
-    input_data_dict = {"x": fake_input}
-    output_data_dict = test_suite.get_truth_val(input_data_dict, device="cpu")
-
-    del test_suite  # destroy class to save memory
-
-    test_suite2 = InferenceTest()
-    test_suite2.load_config(model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams")
-    test_suite2.mkldnn_test(input_data_dict, output_data_dict, delta=1e-2)
-
-    del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.gpu_more_bz_precision
+@pytest.mark.gpu
 def test_gpu_more_bz():
     """
     compared trt gpu batch_size=1-10 resnet50 outputs with true val
     """
     check_model_exist()
 
-    file_path = "./pcpvt_base"
+    file_path = "./resnet50"
     images_size = 224
     batch_size_pool = [1, 5, 10]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
-        test_suite.load_config(
-            model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams"
-        )
+        test_suite.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
         images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
         fake_input = np.array(images_list[0:batch_size]).astype("float32")
-        input_data_dict = {"x": fake_input}
+        input_data_dict = {"inputs": fake_input}
         output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
 
         del test_suite  # destroy class to save memory
 
         test_suite2 = InferenceTest()
-        test_suite2.load_config(
-            model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams"
-        )
+        test_suite2.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
         test_suite2.gpu_more_bz_test(input_data_dict, output_data_dict, delta=1e-5)
 
         del test_suite2  # destroy class to save memory
 
 
 @pytest.mark.jetson
-@pytest.mark.gpu_more_bz_precision
+@pytest.mark.gpu
 def test_jetson_gpu_more_bz():
     """
     compared trt gpu batch_size=1-10 resnet50 outputs with true val
     """
     check_model_exist()
 
-    file_path = "./pcpvt_base"
+    file_path = "./resnet50"
     images_size = 224
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
-        test_suite.load_config(
-            model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams"
-        )
+        test_suite.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
         images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
         fake_input = np.array(images_list[0:batch_size]).astype("float32")
-        input_data_dict = {"x": fake_input}
+        input_data_dict = {"inputs": fake_input}
         output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
 
         del test_suite  # destroy class to save memory
 
         test_suite2 = InferenceTest()
-        test_suite2.load_config(
-            model_file="./pcpvt_base/inference.pdmodel", params_file="./pcpvt_base/inference.pdiparams"
-        )
+        test_suite2.load_config(model_file="./resnet50/inference.pdmodel", params_file="./resnet50/inference.pdiparams")
         test_suite2.gpu_more_bz_test(input_data_dict, output_data_dict, delta=1e-5)
 
         del test_suite2  # destroy class to save memory
