+++ inference/python_api_test/test_class_model/test_swin_transformer_gpu.py
@@ -65,38 +65,7 @@ def test_disable_gpu():
 
 
 @pytest.mark.server
-@pytest.mark.mkldnn_bz1_precision
-def test_mkldnn():
-    """
-    compared mkldnn swin_transformer outputs with true val
-    """
-    check_model_exist()
-    file_path = "./swin_transformer"
-    images_size = 384
-    batch_size_pool = [1]
-    for batch_size in batch_size_pool:
-        test_suite = InferenceTest()
-        test_suite.load_config(
-            model_file="./swin_transformer/inference.pdmodel", params_file="./swin_transformer/inference.pdiparams"
-        )
-        images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-        fake_input = np.array(images_list[0:batch_size]).astype("float32")
-        input_data_dict = {"x": fake_input}
-        output_data_dict = test_suite.get_truth_val(input_data_dict, device="cpu")
-
-        del test_suite  # destroy class to save memory
-
-        test_suite2 = InferenceTest()
-        test_suite2.load_config(
-            model_file="./swin_transformer/inference.pdmodel", params_file="./swin_transformer/inference.pdiparams"
-        )
-        test_suite2.mkldnn_test(input_data_dict, output_data_dict, repeat=1, delta=1e-4)
-
-        del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.mkldnn_bz1_precision
+@pytest.mark.gpu
 def test_gpu_more_bz():
     """
     compared mkldnn swin_transformer outputs with true val
@@ -126,17 +95,16 @@ def test_gpu_more_bz():
         del test_suite2  # destroy class to save memory
 
 
-@pytest.mark.server
 @pytest.mark.jetson
-@pytest.mark.mkldnn_bz1_precision
-def test_gpu_more_bz():
+@pytest.mark.gpu
+def test_jetson_gpu_more_bz():
     """
     compared mkldnn swin_transformer outputs with true val
     """
     check_model_exist()
     file_path = "./swin_transformer"
     images_size = 384
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(
