+++ inference/python_api_test/test_class_model/test_tnt_small_trt_fp16.py
@@ -47,59 +47,10 @@ def test_config():
 
 
 @pytest.mark.server
-@pytest.mark.config_disablegpu_memory
-def test_disable_gpu():
+@pytest.mark.trt_fp16
+def test_trt_fp16_more_bz():
     """
-    test no gpu resources occupied after disable gpu
-    """
-    check_model_exist()
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
-    batch_size = 1
-    fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
-    input_data_dict = {"x": fake_input}
-    test_suite.disable_gpu_test(input_data_dict)
-
-
-@pytest.mark.server
-@pytest.mark.jetson
-@pytest.mark.gpu_more_bz_precision
-def test_gpu_more_bz():
-    """
-    compared gpu batch_size=1,5,10 TNT_small outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./TNT_small"
-    images_size = 224
-    batch_size_pool = [1, 5]
-    for batch_size in batch_size_pool:
-        test_suite = InferenceTest()
-        test_suite.load_config(
-            model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams"
-        )
-        images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-        fake_input = np.array(images_list[0:batch_size]).astype("float32")
-        input_data_dict = {"x": fake_input}
-        output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
-
-        del test_suite  # destroy class to save memory
-
-        test_suite2 = InferenceTest()
-        test_suite2.load_config(
-            model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams"
-        )
-        test_suite2.gpu_more_bz_test(input_data_dict, output_data_dict)
-
-        del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.jetson
-@pytest.mark.trt_fp32_more_bz_precision
-def test_trt_fp32_more_bz():
-    """
-    compared trt fp32 batch_size=1-10 TNT_small outputs with true val
+    compared trt fp16 batch_size=1-10 TNT_small outputs with true val
     """
     check_model_exist()
 
@@ -123,43 +74,15 @@ def test_trt_fp32_more_bz():
             model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams"
         )
         test_suite2.trt_more_bz_test(
-            input_data_dict, output_data_dict, max_batch_size=10, min_subgraph_size=30, precision="trt_fp32"
+            input_data_dict, output_data_dict, delta=1e-2, max_batch_size=10, min_subgraph_size=30, precision="trt_fp16"
         )
 
         del test_suite2  # destroy class to save memory
 
 
-@pytest.mark.server
-@pytest.mark.trt_fp32_multi_thread_bz1_precision
-def test_trt_fp32_bz1_multi_thread():
-    """
-    compared trt fp32 batch_size=1 TNT_small multi_thread outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./TNT_small"
-    images_size = 224
-    batch_size = 1
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
-    images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-    fake_input = np.array(images_list[0:batch_size]).astype("float32")
-    input_data_dict = {"x": fake_input}
-    output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
-
-    del test_suite  # destroy class to save memory
-
-    test_suite2 = InferenceTest()
-    test_suite2.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
-    test_suite2.trt_bz1_multi_thread_test(input_data_dict, output_data_dict, min_subgraph_size=30, precision="trt_fp32")
-
-    del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
 @pytest.mark.jetson
-@pytest.mark.trt_fp16_more_bz_precision
-def test_trt_fp16_more_bz():
+@pytest.mark.trt_fp16
+def test_jetson_trt_fp16_more_bz():
     """
     compared trt fp16 batch_size=1-10 TNT_small outputs with true val
     """
@@ -167,7 +90,7 @@ def test_trt_fp16_more_bz():
 
     file_path = "./TNT_small"
     images_size = 224
-    batch_size_pool = [1, 5]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(
@@ -214,32 +137,5 @@ def test_trt_fp16_bz1_multi_thread():
     test_suite2 = InferenceTest()
     test_suite2.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
     test_suite2.trt_bz1_multi_thread_test(input_data_dict, output_data_dict, min_subgraph_size=30, precision="trt_fp16")
-
-    del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.mkldnn_bz1_precision
-def test_mkldnn():
-    """
-    compared mkldnn TNT_small outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./TNT_small"
-    images_size = 224
-    batch_size = 1
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
-    images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-    fake_input = np.array(images_list[0:batch_size]).astype("float32")
-    input_data_dict = {"x": fake_input}
-    output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
-
-    del test_suite  # destroy class to save memory
-
-    test_suite2 = InferenceTest()
-    test_suite2.load_config(model_file="./TNT_small/inference.pdmodel", params_file="./TNT_small/inference.pdiparams")
-    test_suite2.mkldnn_test(input_data_dict, output_data_dict)
 
     del test_suite2  # destroy class to save memory
