+++ inference/python_api_test/test_class_model/test_vgg11_gpu.py
@@ -46,16 +46,32 @@ def test_config():
 
 
 @pytest.mark.server
-@pytest.mark.trt_fp16
-def test_trtfp16_more_bz():
+@pytest.mark.config_disablegpu_memory
+def test_disable_gpu():
     """
-    compared trt fp16 batch_size=1-10 vgg11 outputs with true val
+    test no gpu resources occupied after disable gpu
+    """
+    check_model_exist()
+    test_suite = InferenceTest()
+    test_suite.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
+    batch_size = 1
+    fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
+    print(fake_input.shape)
+    input_data_dict = {"x": fake_input}
+    test_suite.disable_gpu_test(input_data_dict)
+
+
+@pytest.mark.server
+@pytest.mark.gpu
+def test_gpu_more_bz():
+    """
+    compared gpu vgg11 outputs with true val
     """
     check_model_exist()
 
     file_path = "./vgg11"
     images_size = 224
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1, 5, 10]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
@@ -68,24 +84,16 @@ def test_trtfp16_more_bz():
 
         test_suite2 = InferenceTest()
         test_suite2.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
-        test_suite2.trt_more_bz_test(
-            input_data_dict,
-            output_data_dict,
-            repeat=1,
-            delta=1e-3,
-            gpu_mem=3000,
-            max_batch_size=batch_size,
-            precision="trt_fp16",
-        )
+        test_suite2.gpu_more_bz_test(input_data_dict, output_data_dict)
 
         del test_suite2  # destroy class to save memory
 
 
 @pytest.mark.jetson
-@pytest.mark.trt_fp16
-def test_trtfp16_more_bz():
+@pytest.mark.gpu
+def test_jetson_gpu_more_bz():
     """
-    compared trt fp16 batch_size=1-10 vgg11 outputs with true val
+    compared gpu vgg11 outputs with true val
     """
     check_model_exist()
 
@@ -104,41 +112,6 @@ def test_trtfp16_more_bz():
 
         test_suite2 = InferenceTest()
         test_suite2.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
-        test_suite2.trt_more_bz_test(
-            input_data_dict,
-            output_data_dict,
-            repeat=1,
-            delta=1e-3,
-            gpu_mem=3000,
-            max_batch_size=batch_size,
-            precision="trt_fp16",
-        )
+        test_suite2.gpu_more_bz_test(input_data_dict, output_data_dict)
 
         del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.trt_fp16_multi_thread
-def test_trtfp16_bz1_multi_thread():
-    """
-    compared trt fp16 batch_size=1 vgg11 multi_thread outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./vgg11"
-    images_size = 224
-    batch_size = 1
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
-    images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-    fake_input = np.array(images_list[0:batch_size]).astype("float32")
-    input_data_dict = {"x": fake_input}
-    output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
-
-    del test_suite  # destroy class to save memory
-
-    test_suite2 = InferenceTest()
-    test_suite2.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
-    test_suite2.trt_bz1_multi_thread_test(input_data_dict, output_data_dict, gpu_mem=3000, precision="trt_fp16")
-
-    del test_suite2  # destroy class to save memory
