+++ inference/python_api_test/test_class_model/test_vgg11_trt_fp32.py
@@ -46,22 +46,7 @@ def test_config():
 
 
 @pytest.mark.server
-@pytest.mark.config_disablegpu_memory
-def test_disable_gpu():
-    """
-    test no gpu resources occupied after disable gpu
-    """
-    check_model_exist()
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
-    batch_size = 1
-    fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
-    input_data_dict = {"x": fake_input}
-    test_suite.disable_gpu_test(input_data_dict)
-
-
-@pytest.mark.server
-@pytest.mark.trt_fp32_more_bz_precision
+@pytest.mark.trt_fp32
 def test_trt_fp32_more_bz():
     """
     compared trt fp32 batch_size=1-10 vgg11 outputs with true val
@@ -90,7 +75,7 @@ def test_trt_fp32_more_bz():
 
 @pytest.mark.server
 @pytest.mark.jetson
-@pytest.mark.trt_fp32_more_bz_precision
+@pytest.mark.trt_fp32
 def test_jetson_trt_fp32_more_bz():
     """
     compared trt fp32 batch_size=1-10 vgg11 outputs with true val
@@ -99,7 +84,7 @@ def test_jetson_trt_fp32_more_bz():
 
     file_path = "./vgg11"
     images_size = 224
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(model_file="./vgg11/inference.pdmodel", params_file="./vgg11/inference.pdiparams")
@@ -118,7 +103,7 @@ def test_jetson_trt_fp32_more_bz():
 
 
 @pytest.mark.server
-@pytest.mark.trt_fp32_multi_thread_bz1_precision
+@pytest.mark.trt_fp32_multi_thread
 def test_trtfp32_bz1_multi_thread():
     """
     compared trt fp32 batch_size=1 vgg11 multi_thread outputs with true val
