+++ inference/python_api_test/test_det_model/test_yolov3_trt_fp16.py
@@ -46,37 +46,16 @@ def test_config():
 
 
 @pytest.mark.server
-@pytest.mark.config_disablegpu_memory
-def test_disable_gpu():
+@pytest.mark.trt_fp16_multi_thread
+def test_trt_fp16_more_bz_multi_thread():
     """
-    test no gpu resources occupied after disable gpu
-    """
-    check_model_exist()
-    test_suite = InferenceTest()
-    test_suite.load_config(model_file="./yolov3/model.pdmodel", params_file="./yolov3/model.pdiparams")
-    batch_size = 1
-    im_size = 608
-    im = np.random.randn(batch_size, 3, 608, 608).astype("float32")
-    data = np.random.randn(batch_size, 3, 608, 608).astype("float32")
-    scale_factor = (
-        np.array([im_size * 1.0 / im.shape[0], im_size * 1.0 / im.shape[1]]).reshape((1, 2)).astype(np.float32)
-    )
-    im_shape = np.array([im_size, im_size]).reshape((1, 2)).astype(np.float32)
-    input_data_dict = {"im_shape": im_shape, "image": data, "scale_factor": scale_factor}
-    test_suite.disable_gpu_test(input_data_dict)
-
-
-@pytest.mark.server
-@pytest.mark.mkldnn_more_bz_precision
-def test_more_bz_mkldnn():
-    """
-    compared mkldnn yolov3 batch size = [1,4,8,10] outputs with true val
+    compared trt fp32 batch_size=4 yolov3 multi_thread outputs with true val
     """
     check_model_exist()
 
     file_path = "./yolov3"
     images_size = 608
-    batch_size_pool = [1, 4, 8, 10]
+    batch_size_pool = [4]
     for batch_size in batch_size_pool:
 
         test_suite = InferenceTest()
@@ -113,20 +92,24 @@ def test_more_bz_mkldnn():
 
         output_data_dict = {"save_infer_model/scale_0.tmp_1": scale_0, "save_infer_model/scale_1.tmp_1": scale_1}
         test_suite.load_config(model_file="./yolov3/model.pdmodel", params_file="./yolov3/model.pdiparams")
-        test_suite.mkldnn_test(input_data_dict, output_data_dict, repeat=1, delta=1e-4)
+        test_suite.trt_bz1_multi_thread_test(
+            input_data_dict, output_data_dict, repeat=1, delta=1e-4, precision="trt_fp16"
+        )
+
+        del test_suite  # destroy class to save memory
 
 
-@pytest.mark.p1
-@pytest.mark.gpu_more_bz_precision
-def test_gpu_more_bz():
+@pytest.mark.server
+@pytest.mark.trt_fp16_more
+def test_trt_fp16_more_bz():
     """
-    compared gpu yolov3 batch size = [1,4,8] outputs with true val
+    compared trt fp32 batch_size=1,5,10 yolov3 outputs with true val
     """
     check_model_exist()
 
     file_path = "./yolov3"
     images_size = 608
-    batch_size_pool = [1, 4, 8]
+    batch_size_pool = [1, 5, 10]
     for batch_size in batch_size_pool:
 
         test_suite = InferenceTest()
@@ -163,21 +146,22 @@ def test_gpu_more_bz():
 
         output_data_dict = {"save_infer_model/scale_0.tmp_1": scale_0, "save_infer_model/scale_1.tmp_1": scale_1}
         test_suite.load_config(model_file="./yolov3/model.pdmodel", params_file="./yolov3/model.pdiparams")
-        test_suite.gpu_more_bz_test(input_data_dict, output_data_dict, repeat=1, delta=1e-4)
+        test_suite.trt_more_bz_test(input_data_dict, output_data_dict, repeat=1, delta=1e-4, precision="trt_fp16")
+
+        del test_suite  # destroy class to save memory
 
 
-@pytest.mark.server
 @pytest.mark.jetson
-@pytest.mark.gpu_more_bz_precision
-def test_gpu_more_bz():
+@pytest.mark.trt_fp16
+def test_jetson_trt_fp16_more_bz():
     """
-    compared gpu yolov3 batch size = [1,4,8] outputs with true val
+    compared trt fp32 batch_size=1,5,10 yolov3 outputs with true val
     """
     check_model_exist()
 
     file_path = "./yolov3"
     images_size = 608
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
 
         test_suite = InferenceTest()
@@ -214,4 +198,8 @@ def test_gpu_more_bz():
 
         output_data_dict = {"save_infer_model/scale_0.tmp_1": scale_0, "save_infer_model/scale_1.tmp_1": scale_1}
         test_suite.load_config(model_file="./yolov3/model.pdmodel", params_file="./yolov3/model.pdiparams")
-        test_suite.gpu_more_bz_test(input_data_dict, output_data_dict, repeat=1, delta=1e-4)
+        test_suite.trt_more_bz_test(
+            input_data_dict, output_data_dict, repeat=1, max_batch_size=10, delta=6e-2, precision="trt_fp16"
+        )
+
+        del test_suite  # destroy class to save memory
