+++ inference/python_api_test/test_ocr_model/test_ocr_det_mv3_db_gpu.py
@@ -66,38 +66,7 @@ def test_disable_gpu():
 
 
 @pytest.mark.server
-@pytest.mark.mkldnn_bz1_precision
-def test_mkldnn():
-    """
-    compared mkldnn ocr_det_mv3_db outputs with true val
-    """
-    check_model_exist()
-
-    file_path = "./ocr_det_mv3_db"
-    images_size = 224
-    batch_size = 1
-    test_suite = InferenceTest()
-    test_suite.load_config(
-        model_file="./ocr_det_mv3_db/inference.pdmodel", params_file="./ocr_det_mv3_db/inference.pdiparams"
-    )
-    images_list, npy_list = test_suite.get_images_npy(file_path, images_size)
-    fake_input = np.array(images_list[0:batch_size]).astype("float32")
-    input_data_dict = {"x": fake_input}
-    output_data_dict = test_suite.get_truth_val(input_data_dict, device="gpu")
-
-    del test_suite  # destroy class to save memory
-
-    test_suite2 = InferenceTest()
-    test_suite2.load_config(
-        model_file="./ocr_det_mv3_db/inference.pdmodel", params_file="./ocr_det_mv3_db/inference.pdiparams"
-    )
-    test_suite2.mkldnn_test(input_data_dict, output_data_dict, delta=1e-4)
-
-    del test_suite2  # destroy class to save memory
-
-
-@pytest.mark.server
-@pytest.mark.gpu_more_bz_precision
+@pytest.mark.gpu_more
 def test_gpu_more_bz():
     """
     compared trt fp32 batch_size=1,2 ocr_det_mv3_db outputs with true val
@@ -130,7 +99,7 @@ def test_gpu_more_bz():
 
 
 @pytest.mark.jetson
-@pytest.mark.gpu_more_bz_precision
+@pytest.mark.gpu_more
 def test_jetson_gpu_more_bz():
     """
     compared trt fp32 more batch_size ocr_det_mv3_db outputs with true val
@@ -139,7 +108,7 @@ def test_jetson_gpu_more_bz():
 
     file_path = "./ocr_det_mv3_db"
     images_size = 640
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     for batch_size in batch_size_pool:
         test_suite = InferenceTest()
         test_suite.load_config(
