+++ inference/python_api_test/test_ocr_model/test_ocr_det_mv3_db_trt_fp32.py
@@ -48,24 +48,7 @@ def test_config():
 
 
 @pytest.mark.server
-@pytest.mark.config_disablegpu_memory
-def test_disable_gpu():
-    """
-    test no gpu resources occupied after disable gpu
-    """
-    check_model_exist()
-    test_suite = InferenceTest()
-    test_suite.load_config(
-        model_file="./ocr_det_mv3_db/inference.pdmodel", params_file="./ocr_det_mv3_db/inference.pdiparams"
-    )
-    batch_size = 1
-    fake_input = np.random.randn(batch_size, 3, 224, 224).astype("float32")
-    input_data_dict = {"x": fake_input}
-    test_suite.disable_gpu_test(input_data_dict)
-
-
-@pytest.mark.server
-@pytest.mark.trt_fp32_multi_thread_bz1_dynamic_precision
+@pytest.mark.trt_fp32_multi_thread
 def test_trt_fp32_bz1_dynamic_multi_thread():
     """
     compared trt fp32 batch_size=1 ocr_det_mv3_db multi_thread outputs with true val
@@ -167,7 +150,7 @@ def test_trt_fp32_bz1_dynamic_multi_thread():
 
 
 @pytest.mark.server
-@pytest.mark.trt_fp32_more_bz_dynamic_precision
+@pytest.mark.trt_fp32
 def test_trtfp32_more_bz_dynamic_bz():
     """
     compared trt fp32 batch_size=1,2 ocr_det_mv3_db outputs with true val
@@ -269,9 +252,8 @@ def test_trtfp32_more_bz_dynamic_bz():
         del test_suite2  # destroy class to save memory
 
 
-@pytest.mark.server
 @pytest.mark.jetson
-@pytest.mark.trt_fp32_more_bz_dynamic_precision
+@pytest.mark.trt_fp32
 def test_jetson_trtfp32_more_bz_dynamic_bz():
     """
     compared trt fp32 batch_size=1,2 ocr_det_mv3_db outputs with true val
@@ -280,7 +262,7 @@ def test_jetson_trtfp32_more_bz_dynamic_bz():
 
     file_path = "./ocr_det_mv3_db"
     images_size = 640
-    batch_size_pool = [1, 2]
+    batch_size_pool = [1]
     max_batch_size = 5
     names = [
         "x",
